responseBody 구성 변경 #2

+++ src/main/java/com/web/dto/EventPageDto.java
@@ -2,12 +2,14 @@ package com.web.dto;
 
 import java.util.List;
 
+import com.web.domain.CafeEvent;
+
 import lombok.Data;
 
 @Data
 public class EventPageDto {
 	private List<EventDto> content;
-	private int count;
+	private Long count;
 	private int size;
 	private int page;
